Create README.md
Looking for an opportunity to utilize my acquired skills and training to help the company and my future peers grow. I want my efforts to make a considerable difference to the company and help in their consequent success.”
 Interested in the position of Web Developer in Mass Markets and strong familiarity with JavaScript (object-oriented), HTML5, and CSS3 with excellent communication and interpersonal skills.
Detail-oriented individual with a BE in Mechanical Engineering searching for a suitable role at an established organization to utilize my knowledge and skills while contributing to the growth of the company.
Looking forward to an entry-level position to kickstart my career. I am adaptable and a quick learner. I can assure you of my professional and personal growth because I have what it takes to forge a successful path for both the company and my future.”